Resolve sending dublicate data
